IndexedDocumentDaoImplTest - fixes

+++ src/main/java/com/glueball/snoop/dao/setter/IndexedDocumentBatchInsertSetter.java
@@ -61,7 +61,7 @@ public final class IndexedDocumentBatchInsertSetter implements
         pstmt.setTimestamp(index++, docs.get(i).getLastIndexedTime());
         pstmt.setString(index++, docs.get(i).getContentType());
         pstmt.setString(index++, docs.get(i).getIndexState());
-        pstmt.setInt(index++, docs.get(i).getLock());
+        pstmt.setLong(index++, docs.get(i).getLock());
 
         if (docs.get(i).getLockTime() != null) {
             pstmt.setTimestamp(index++, docs.get(i).getLockTime());
